moves ListenerService to its own file
